Add friend / remove friend buttons are properly working on other user's profile

+++ backend/src/friend/friend.service.ts
@@ -120,6 +120,8 @@ export class FriendService{
     // Deleting friendship
     async deleteFriendShip(userId: number, friendId: number)
     {
+		if (await this.existingFriendship(userId, friendId) == false)
+			return;
         const deleteFriend = await this.prisma.friend.delete({
             where: {
                 user_id_friend_id: {
